# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "description": "Dokumentasi API User menggunakan Swagger di Laravel",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server API"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Mendapatkan daftar semua pengguna",
                "operationId": "eaedf6aa84fde38a08bf5d87308169f7",
                "responses": {
                    "200": {
                        "description": "Daftar pengguna",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Menambahkan pengguna baru",
                "operationId": "e2308aae8643f8f4d79821a6399114f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil menambahkan data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Data yang diberikan tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreUpdateValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Mendapatkan pengguna berdasarkan ID",
                "operationId": "e0e2899dacd82633162b668cc31d8116",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pengguna",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengguna berdasarkan ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataNotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Menambahkan pengguna baru",
                "operationId": "15f5757522724ebca9bac30479321b45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pengguna",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil mengubah data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Data yang diberikan tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreUpdateValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataNotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Menghapus pengguna berdasarkan ID",
                "operationId": "98b6e09e3d231384330198299feb017f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pengguna",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengguna berdasarkan ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DataNotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreUpdateValidationError": {
                "title": "Validation Error",
                "description": "Schema untuk response validasi gagal",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 422
                    },
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data yang diberikan tidak valid"
                    },
                    "errors": {
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Nama tidak boleh kosong."
                                }
                            },
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Email tidak boleh kosong."
                                }
                            },
                            "age": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Usia tidak boleh kosong."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "DataNotFoundError": {
                "title": "Data Not Found Error",
                "description": "Schema untuk response validasi gagal",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 404
                    },
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data tidak ditemukan"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "required": [
                    "name",
                    "email",
                    "age"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Faiz Rahmat Hidayat"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "faizrahmat.hidayat06@gmail.com"
                    },
                    "age": {
                        "type": "integer",
                        "example": 30
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "age"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Faiz Rahmat"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "faizrahmat.hidayat06@gmail.com"
                    },
                    "age": {
                        "type": "integer",
                        "example": 30
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}